Cast numMacrVar from size_t to int

+++ Src/model.cpp
@@ -319,7 +319,7 @@ void DefineMacroVars(std::vector<VariableTypes> types,
         assert(components.size() == 0);
     }
 
-    int numMacroVar{names.size()};
+    int numMacroVar{static_cast<int>(names.size())};
 
     if (numMacroVar > 0) {
         ops_printf("There are %i macroscopic variables defined.\n",
